argus/argus-dev/requirements.txt
astroid==2.2.5
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.2
pylint==2.3.1
PyYAML==5.1
scapy==2.4.2
scipy==1.2.1
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
